pyproject.toml
[project]
name = "impactracer"
version = "4.0.0"
description = "Change Impact Analysis via RAG with multilingual embeddings and structural graph"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
    "tree-sitter==0.21.3",
    "tree-sitter-languages>=1.10.2",
    "mistune>=3.0",
    "FlagEmbedding>=1.2.5",
    "transformers>=4.44.2,<5.0",
    "google-genai>=0.3.0",
    "chromadb>=0.5.0",
    "rank_bm25>=0.2.2",
    "networkx>=3.2",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "typer[all]>=0.12",
    "loguru>=0.7",
    "python-dotenv>=1.0",
    "tiktoken>=0.7",
    "scipy>=1.13",
    "pandas>=2.2",
    "numpy>=1.26",
    "huggingface-hub>=0.23",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.11",
]

[project.scripts]
impactracer = "impactracer.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["impactracer"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
